feat(tests): add comprehensive integration tests for PostgreSQL todo store
Add integration test suite for pgtodo store using testcontainers with
PostgreSQL database to validate CRUD operations in realistic
environment.

**Features Added:** - SetUpTestDB function with testcontainers
PostgreSQL setup - Database migration execution in test environment -
Integration tests covering Create, List, and Get operations - Proper
test isolation with containerized database

**Refactoring Required:** - Fixed todo title numbering inconsistency in
test data generation - Changed from "test0, test1, test2" to "test1,
test2, test3" naming - Updated test expectations to match ID-to-title
mapping (ID 1 → "test1") - Resolved test failure caused by title
mismatch in GetTodo validation

The integration tests provide confidence in store functionality against
a real PostgreSQL database, complementing existing unit tests with
end-to-end validation of database interactions.

+++ domain/web_dto.go
@@ -5,7 +5,7 @@ package domain
 // Similar to a Java DTO class or a JS object used in APIs.
 
 type TodoDTO struct {
-	ID        int    `json:"id"` // json tag for easy JSON encoding (like @JsonProperty in Java)
+	ID        int64  `json:"id"` // json tag for easy JSON encoding (like @JsonProperty in Java)
 	Title     string `json:"title"`
 	Done      bool   `json:"done"`
 	CreatedAt string `json:"createdAt"`
